creating helper file and moving bitfield/bytefileds definitions to this file
